src/main/java/ru/ssau/tk/DDRyskovaCo/practice/Collection/Route.java
package ru.ssau.tk.DDRyskovaCo.practice.Collection;

import java.util.*;

public class Route implements Iterable<Location> {

    private final List<Location> locations = new ArrayList<>();

    public List<Location> getLocations() {
        return locations;
    }

    public void addLocation(Location location) {
        this.locations.add(location);
    }

    public void addLocationByIndex(int index, Location location) {
        this.locations.add(index, location);
    }

    public void removeLocation(int index) {
        this.locations.remove(index);
    }

    public Location getFirstLocation() {
        return locations.get(0);
    }

    public Location getLastLocation() {
        return locations.get(locations.size() - 1);
    }

    @Override
    public Iterator<Location> iterator() {
        return locations.iterator();
    }
}
